Reinforcement Learning Project -2
In project 1, you may realize that when you have a large grid world maze setup, it takes a long
time for the agent to learn a value table. One way to eliminate this challenge is to use neural
networks to approximate the value function. There are two options provided below and you
may choose either one to implement.
A. Based on your results in project 1, you can choose to build a neural network (or deep
neural network) to approximate your obtained Q or V table.
B. You can design another complex grid world example and develop the Q-learning (or
deep Q-learning) method based on that.
Either way, you are using a neural network to generate your Q or V value so that you can guide
the agent to move to achieve the goal.